@@ -13,13 +14,13 @@
     "Dário Vieira <dario.junior3@gmail.com>"
   ],
   "engines": {
-    "node": "=14.19.3"
+    "node": "=19.1.0"
   },
   "license": "AGPL-3.0",
   "preferGlobal": true,
   "private": false,
   "bin": {
-    "gbot": "./boot.js"
+    "gbot": "./boot.cjs"
   },
   "repository": {
     "type": "git",
@@ -49,131 +50,129 @@
     "commit": "git-cz"
   },
   "dependencies": {
-    "@azure/cognitiveservices-computervision": "8.1.0",
-    "@azure/ms-rest-js": "2.5.1",
-    "@google-cloud/pubsub": "2.13.0",
-    "@google-cloud/translate": "6.2.6",
-    "@hubspot/api-client": "5.0.0",
-    "@microsoft/microsoft-graph-client": "2.2.1",
-    "@semantic-release/changelog": "5.0.1",
-    "@semantic-release/exec": "5.0.0",
-    "@semantic-release/git": "9.0.0",
-    "@sendgrid/mail": "7.4.4",
-    "@types/node": "8.0.0",
-    "@types/validator": "13.1.4",
-    "adal-node": "0.2.2",
-    "adm-zip": "0.5.6",
-    "alasql": "^1.7.3",
+    "@azure/cognitiveservices-computervision": "8.2.0",
+    "@azure/keyvault-keys": "4.6.0",
+    "@azure/ms-rest-js": "2.6.2",
+    "@azure/msal-node": "1.14.3",
+    "@azure/search-documents": "^11.3.1",
+    "@google-cloud/pubsub": "3.2.1",
+    "@google-cloud/translate": "7.0.4",
+    "@hubspot/api-client": "7.1.2",
+    "@microsoft/microsoft-graph-client": "3.0.4",
+    "@nosferatu500/textract": "3.1.2",
+    "@semantic-release/changelog": "6.0.1",
+    "@semantic-release/exec": "6.0.3",
+    "@semantic-release/git": "10.0.1",
+    "@sendgrid/mail": "7.7.0",
+    "@types/node": "18.11.9",
+    "@types/validator": "13.7.10",
+    "adm-zip": "0.5.9",
+    "alasql": "2.1.6",
     "any-shell-escape": "0.1.1",
     "async-promises": "0.2.3",
-    "azure-arm-cognitiveservices": "3.0.0",
-    "azure-arm-resource": "7.4.0",
-    "azure-arm-search": "1.3.0-preview",
-    "azure-arm-sql": "5.7.0",
-    "azure-arm-website": "5.7.0",
-    "azure-search-client": "3.1.5",
     "basic-auth": "2.0.1",
-    "billboard.js": "^3.4.1",
+    "billboard.js": "3.6.3",
     "bluebird": "3.7.2",
-    "body-parser": "1.19.0",
-    "botbuilder": "4.11.0",
-    "botbuilder-adapter-facebook": "1.0.11",
-    "botbuilder-ai": "4.11.0",
-    "botbuilder-dialogs": "4.11.0",
-    "botframework-connector": "4.11.0",
-    "botlib": "1.10.9",
-    "c3-chart-maker": "^0.2.8",
-    "cli-progress": "^3.11.2",
+    "body-parser": "1.20.1",
+    "botbuilder": "4.18.0",
+    "botbuilder-adapter-facebook": "1.0.12",
+    "botbuilder-ai": "4.18.0",
+    "botbuilder-dialogs": "4.18.0",
+    "botframework-connector": "4.18.0",
+    "botlib": "3.0.0",
+    "c3-chart-maker": "0.2.8",
+    "chrome-remote-interface": "0.31.3",
+    "cli-progress": "3.11.2",
     "cli-spinner": "0.2.10",
-    "core-js": "3.14.0",
-    "data-forge": "^1.9.5",
-    "date-diff": "0.2.2",
-    "docxtemplater": "^3.31.1",
+    "core-js": "3.26.1",
+    "data-forge": "1.9.6",
+    "date-diff": "1.0.2",
+    "docxtemplater": "3.32.4",
     "dotenv-extended": "2.9.0",
-    "exceljs": "4.2.1",
-    "express": "4.17.1",
+    "exceljs": "4.3.0",
+    "express": "4.18.2",
     "express-remove-route": "1.0.0",
-    "ffmpeg-static": "4.3.0",
-    "google-libphonenumber": "3.2.21",
-    "googleapis": "75.0.0",
-    "ibm-watson": "6.1.1",
-    "js-beautify": "1.13.13",
-    "keyv": "^4.5.0",
-    "lodash": "^4.17.21",
-    "luxon": "2.0.2",
-    "mammoth": "^1.4.21",
-    "marked": "2.0.7",
-    "momentjs": "2.0.0",
-    "ms-rest-azure": "3.0.0",
+    "ffmpeg-static": "5.1.0",
+    "google-libphonenumber": "3.2.31",
+    "googleapis": "109.0.1",
+    "ibm-watson": "7.1.2",
+    "indent.js": "0.3.5",
+    "js-beautify": "1.14.7",
+    "keyv": "4.5.2",
+    "koa": "2.13.4",
+    "koa-body": "6.0.1",
+    "koa-router": "12.0.0",
+    "lodash": "4.17.21",
+    "luxon": "3.1.0",
+    "mammoth": "1.5.1",
+    "marked": "4.2.2",
+    "moment": "1.3.0",
+    "ms-rest-azure": "3.0.2",
     "nexmo": "2.9.1",
-    "node-cron": "3.0.0",
-    "node-nlp": "^3.10.2",
-    "node-tesseract-ocr": "^2.2.1",
-    "npm": "7.21.0",
-    "opn": "6.0.0",
+    "node-cron": "3.0.2",
+    "node-nlp": "^4.24.0",
+    "node-tesseract-ocr": "2.2.1",
+    "npm": "9.1.2",
+    "open": "8.4.0",
     "pdf-extraction": "1.0.2",
-    "pdfkit": "^0.13.0",
-    "phone": "2.4.21",
-    "pizzip": "^3.0.6",
+    "pdfkit": "0.13.0",
+    "phone": "3.1.30",
+    "pizzip": "3.1.3",
     "pptxtemplater": "1.0.5",
     "pragmatismo-io-framework": "1.0.20",
-    "prism-media": "1.3.1",
-    "public-ip": "4.0.4",
-    "puppeteer": "13.7.0",
-    "puppeteer-extra": "^3.3.4",
-    "puppeteer-extra-plugin-stealth": "2.4.5",
-    "qrcode": "^1.5.0",
+    "prism-media": "1.3.4",
+    "public-ip": "6.0.1",
+    "punycode": "2.1.1",
+    "puppeteer": "19.2.2",
+    "puppeteer-extra": "3.3.4",
+    "puppeteer-extra-plugin-stealth": "2.11.1",
+    "qrcode": "1.5.1",
     "qrcode-terminal": "0.12.0",
     "readline": "1.3.0",
     "reflect-metadata": "0.1.13",
-    "request-promise": "4.2.5",
-    "request-promise-native": "1.0.8",
     "rimraf": "3.0.2",
     "safe-buffer": "5.2.1",
     "scanf": "1.1.2",
-    "sequelize": "6.5.0",
-    "sequelize-cli": "6.2.0",
-    "sequelize-typescript": "2.1.0",
-    "simple-git": "2.39.1",
+    "sequelize": "6.25.7",
+    "sequelize-cli": "6.5.2",
+    "sequelize-typescript": "2.1.5",
+    "simple-git": "3.15.0",
     "speakingurl": "14.0.1",
-    "sppull": "2.7.0",
     "ssr-for-bots": "1.0.1-c",
     "strict-password-generator": "1.1.2",
-    "swagger-client": "2.1.18",
-    "tabulator-tables": "^5.2.6",
-    "tedious": "14.0.0",
-    "textract": "2.5.0",
-    "twitter-api-v2": "1.12.7",
-    "typescript": "3.6.4",
-    "url-join": "4.0.1",
+    "swagger-client": "^3.18.5",
+    "tabulator-tables": "5.4.2",
+    "tedious": "15.1.2",
+    "twitter-api-v2": "1.12.9",
+    "typescript": "4.9.3",
+    "typescript-rest-rpc": "^1.0.7",
+    "url-join": "5.0.0",
     "vbscript-to-typescript": "1.0.8",
-    "vhost": "^3.0.2",
-    "vm2": "^3.9.11",
+    "vhost": "3.0.2",
+    "vm2": "3.9.11",
+    "vm2-process": "2.1.1",
     "walk-promise": "0.2.0",
     "washyourmouthoutwithsoap": "1.0.2",
-    "whatsapp-web.js": "github:pedroslopez/whatsapp-web.js#fix-buttons-list",
-    "winston": "^2.4.2",
+    "whatsapp-web.js": "1.18.3",
+    "winston": "3.8.2",
     "winston-logs-display": "^1.0.0",
     "yarn": "^1.22.19"
   },
   "devDependencies": {
-    "@types/puppeteer": "5.4.6",
-    "@types/url-join": "4.0.0",
-    "@types/winston": "2.4.4",
-    "ban-sensitive-files": "1.9.15",
-    "commitizen": "4.2.4",
+    "@types/url-join": "4.0.1",
+    "ban-sensitive-files": "1.9.18",
+    "commitizen": "4.2.5",
     "cz-conventional-changelog": "3.3.0",
     "dependency-check": "4.1.0",
-    "git-issues": "1.3.1",
+    "git-issues": "^1.0.0",
     "license-checker": "25.0.1",
-    "ngrok": "4.0.1",
-    "nsp": "3.2.1",
-    "prettier-standard": "16.4.1",
-    "semantic-release": "17.4.3",
-    "simple-commit-message": "4.1.2",
-    "travis-deploy-once": "5.0.11",
-    "ts-node": "10.0.0",
-    "tslint": "6.1.2"
+    "ngrok": "4.3.3",
+    "prettier-standard": "^15.0.1",
+    "semantic-release": "19.0.5",
+    "simple-commit-message": "^1.1.0",
+    "travis-deploy-once": "^3.0.0",
+    "ts-node": "10.9.1",
+    "tslint": "6.1.3"
   },
   "eslintConfig": {
     "env": {
