@@ -7,27 +7,26 @@
   "license": "AGPL-3.0",
   "homepage": ".",
   "dependencies": {
-    "@midudev/react-static-content": "^1.0.4",
-    "ajv": "^8.6.0",
-    "botframework-directlinejs": "0.14.1",
-    "botframework-webchat": "^4.13.0",
+    "@midudev/react-static-content": "1.0.4",
+    "ajv": "8.11.2",
+    "botframework-directlinejs": "0.15.1",
+    "botframework-webchat": "4.15.5",
     "deep-extend": "0.6.0",
-    "eslint": "7.11.0",
+    "eslint": "8.28.0",
     "fetch": "1.1.0",
-    "msal": "^1.4.11",
-    "powerbi-client": "2.18.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "msal": "1.4.17",
+    "powerbi-client": "2.22.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-helmet": "6.1.0",
-    "react-modern-audio-player": "^1.2.2",
-    "react-player": "^2.9.0",
+    "react-player": "2.11.0",
     "react-powerbi": "0.9.1",
-    "react-scripts": "^4.0.3",
-    "react-super-seo": "^1.0.6",
-    "react-transition-group": "^4.4.2",
-    "rxjs": "^7.1.0",
-    "url-join": "4.0.1",
-    "webpack": "4.44.2"
+    "react-scripts": "5.0.1",
+    "react-super-seo": "1.0.7",
+    "react-transition-group": "4.4.5",
+    "rxjs": "7.5.7",
+    "url-join": "5.0.0",
+    "webpack": "5.75.0"
   },
   "scripts": {
     "start": "react-scripts start",
