@@ -37,10 +37,10 @@
 'use strict';
 
 import { GBDialogStep, GBLog, GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
-const urlJoin = require('url-join');
-import { FeedbackDialog } from './dialogs/FeedbackDialog';
-import { QualityDialog } from './dialogs/QualityDialog';
-import { GuaribasQuestionAlternate } from './models/index';
+import urlJoin from 'url-join';
+import { FeedbackDialog } from './dialogs/FeedbackDialog.js';
+import { QualityDialog } from './dialogs/QualityDialog.js';
+import { GuaribasQuestionAlternate } from './models/index.js';
 
 import { Sequelize } from 'sequelize-typescript';
 
@@ -49,26 +49,26 @@ import { Sequelize } from 'sequelize-typescript';
  */
 export class GBCustomerSatisfactionPackage implements IGBPackage {
   public sysPackages: IGBPackage[];
-  public async getDialogs(min: GBMinInstance) {
+  public async getDialogs (min: GBMinInstance) {
     GBLog.verbose(`getDialogs called.`);
   }
-  public async unloadPackage(core: IGBCoreService): Promise<void> {
+  public async unloadPackage (core: IGBCoreService): Promise<void> {
     GBLog.verbose(`unloadPackage called.`);
   }
-  public async unloadBot(min: GBMinInstance): Promise<void> {
+  public async unloadBot (min: GBMinInstance): Promise<void> {
     GBLog.verbose(`unloadBot called.`);
   }
-  public async onNewSession(min: GBMinInstance, step: GBDialogStep): Promise<void> {
+  public async onNewSession (min: GBMinInstance, step: GBDialogStep): Promise<void> {
     GBLog.verbose(`onNewSession called.`);
   }
-  public async onExchangeData(min: GBMinInstance, kind: string, data: any) {
+  public async onExchangeData (min: GBMinInstance, kind: string, data: any) {
     GBLog.verbose(`onExchangeData called.`);
   }
 
-  public async loadPackage(core: IGBCoreService, sequelize: Sequelize): Promise<void> {
+  public async loadPackage (core: IGBCoreService, sequelize: Sequelize): Promise<void> {
     core.sequelize.addModels([GuaribasQuestionAlternate]);
   }
-  public async loadBot(min: GBMinInstance): Promise<void> {
+  public async loadBot (min: GBMinInstance): Promise<void> {
     FeedbackDialog.setup(min.bot, min);
     QualityDialog.setup(min.bot, min);
   }
