@@ -36,31 +36,23 @@
 
 'use strict';
 
-const urlJoin = require('url-join');
-
-import { BotAdapter, CardFactory, MessageFactory } from 'botbuilder';
-import { WaterfallDialog } from 'botbuilder-dialogs';
 import { GBLog, GBMinInstance, IGBDialog } from 'botlib';
-import { GBAdminService } from '../../admin.gbapp/services/GBAdminService';
-import { GBConversationalService } from '../../core.gbapp/services/GBConversationalService';
-import { Messages } from '../strings';
-const phoneUtil = require('google-libphonenumber').PhoneNumberUtil.getInstance();
-const phone = require('phone');
+import { GBAdminService } from '../../admin.gbapp/services/GBAdminService.js';
+import { Messages } from '../strings.js';
+import * as phone from 'google-libphonenumber';
 
 /**
  * Dialogs for handling Menu control.
  */
 export class ProfileDialog extends IGBDialog {
-
-  public static getNameDialog(min: GBMinInstance) {
-
+  public static getNameDialog (min: GBMinInstance) {
     return {
-      id: '/profile_name', waterfall: [
+      id: '/profile_name',
+      waterfall: [
         async step => {
           if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
             return await step.beginDialog('/auth');
-          }
-          else{
+          } else {
             return await step.next(step.options);
           }
         },
@@ -72,7 +64,7 @@ export class ProfileDialog extends IGBDialog {
         async step => {
           const locale = step.context.activity.locale;
 
-          const fullName = (text) => {
+          const fullName = text => {
             return text.match(/^[a-zA-Z]+(([',. -][a-zA-Z ])?[a-zA-Z]*)*$/gi);
           };
 
@@ -85,21 +77,20 @@ export class ProfileDialog extends IGBDialog {
             step.activeDialog.state.options.name = value[0];
 
             return await step.replaceDialog('/profile_mobile', step.activeDialog.state.options);
-
           }
-        }]
+        }
+      ]
     };
   }
 
-  public static getMobileDialog(min: GBMinInstance) {
-
+  public static getMobileDialog (min: GBMinInstance) {
     return {
-      id: '/profile_mobile', waterfall: [
+      id: '/profile_mobile',
+      waterfall: [
         async step => {
           if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
             return await step.beginDialog('/auth');
-          }
-          else{
+          } else {
             return await step.next(step.options);
           }
         },
@@ -112,14 +103,15 @@ export class ProfileDialog extends IGBDialog {
           const locale = step.context.activity.locale;
           let phoneNumber;
           try {
-            phoneNumber = phone(step.result, 'BRA')[0]; // TODO: Use accordingly to the person.
-            phoneNumber = phoneUtil.parse(phoneNumber);
+            let p = phone.PhoneNumberUtil.getInstance();
+            phoneNumber = p(step.result, 'BRA')[0]; // TODO: Use accordingly to the person.
+            phoneNumber = phone.phoneUtil.parse(phoneNumber);
           } catch (error) {
             await step.context.sendActivity(Messages[locale].validation_enter_valid_mobile);
 
             return await step.replaceDialog('/profile_mobile', step.activeDialog.state.options);
           }
-          if (!phoneUtil.isPossibleNumber(phoneNumber)) {
+          if (!phone.phoneUtil.isPossibleNumber(phoneNumber)) {
             await step.context.sendActivity(Messages[locale].validation_enter_valid_mobile);
 
             return await step.replaceDialog('/profile_mobile', step.activeDialog.state.options);
@@ -129,19 +121,19 @@ export class ProfileDialog extends IGBDialog {
 
             return await step.replaceDialog('/profile_mobile_confirm', step.activeDialog.state.options);
           }
-        }]
+        }
+      ]
     };
   }
 
-  public static getMobileConfirmDialog(min: GBMinInstance) {
-
+  public static getMobileConfirmDialog (min: GBMinInstance) {
     return {
-      id: '/profile_mobile_confirm', waterfall: [
+      id: '/profile_mobile_confirm',
+      waterfall: [
         async step => {
           if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
             return await step.beginDialog('/auth');
-          }
-          else{
+          } else {
             return await step.next(step.options);
           }
         },
@@ -150,8 +142,10 @@ export class ProfileDialog extends IGBDialog {
           const locale = step.context.activity.locale;
           const from = step.activeDialog.state.options.mobile;
           if (min.whatsAppDirectLine) {
-
-            await min.whatsAppDirectLine.sendToDevice(from, `${step.activeDialog.state.options.mobileCode} is your General Bots creation code.`);
+            await min.whatsAppDirectLine.sendToDevice(
+              from,
+              `${step.activeDialog.state.options.mobileCode} is your General Bots creation code.`
+            );
           } else {
             GBLog.info(`WhatsApp not configured. Here is the code: ${step.activeDialog.state.options.mobileCode}.`);
           }
@@ -168,18 +162,19 @@ export class ProfileDialog extends IGBDialog {
           } else {
             await step.replaceDialog('/profile_email', step.activeDialog.state.options);
           }
-        }]
+        }
+      ]
     };
   }
 
-  public static getEmailDialog(min: GBMinInstance) {
+  public static getEmailDialog (min: GBMinInstance) {
     return {
-      id: '/profile_email', waterfall: [
+      id: '/profile_email',
+      waterfall: [
         async step => {
           if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
             return await step.beginDialog('/auth');
-          }
-          else{
+          } else {
             return await step.next(step.options);
           }
         },
@@ -190,7 +185,7 @@ export class ProfileDialog extends IGBDialog {
         async step => {
           const locale = step.context.activity.locale;
 
-          const extractEntity = (text) => {
+          const extractEntity = text => {
             return text.match(/([a-zA-Z0-9._-]+@[a-zA-Z0-9._-]+\.[a-zA-Z0-9_-]+)/gi);
           };
 
@@ -203,7 +198,8 @@ export class ProfileDialog extends IGBDialog {
             step.activeDialog.state.options.email = value[0];
             await step.replaceDialog(`/${step.activeDialog.state.options.nextDialog}`, step.activeDialog.state.options);
           }
-        }]
+        }
+      ]
     };
   }
 }
