@@ -30,35 +30,29 @@
 |                                                                             |
 \*****************************************************************************/
 
-const urlJoin = require('url-join');
-
-const Swagger = require('swagger-client');
-const fs = require('fs');
-const Path = require('path');
+import urlJoin from 'url-join';
+import Swagger from 'swagger-client';
+import Path from 'path';
+import Fs from 'fs';
 import { GBLog, GBMinInstance, GBService, IGBPackage } from 'botlib';
 import { CollectionUtil } from 'pragmatismo-io-framework';
-import * as request from 'request-promise-native';
-import { GBServer } from '../../../src/app';
-import { GBConversationalService } from '../../core.gbapp/services/GBConversationalService';
-import { SecService } from '../../security.gbapp/services/SecService';
-import { Messages } from '../strings';
-import { GuaribasUser } from '../../security.gbapp/models';
-import { DialogKeywords } from '../../basic.gblib/services/DialogKeywords';
-import { GBAdminService } from '../../admin.gbapp/services/GBAdminService';
-import { GBMinService } from '../../core.gbapp/services/GBMinService';
-import { GBConfigService } from '../../core.gbapp/services/GBConfigService';
-
-const puppeteer = require('puppeteer');
-
-const { MessageMedia, Client, LocalAuth } = require('whatsapp-web.js');
-const qrcode = require('qrcode-terminal');
-
+import { GBServer } from '../../../src/app.js';
+import { GBConversationalService } from '../../core.gbapp/services/GBConversationalService.js';
+import { SecService } from '../../security.gbapp/services/SecService.js';
+import { Messages } from '../strings.js';
+import { GuaribasUser } from '../../security.gbapp/models/index.js';
+import { GBMinService } from '../../core.gbapp/services/GBMinService.js';
+import { GBConfigService } from '../../core.gbapp/services/GBConfigService.js';
+import * as wpp from 'whatsapp-web.js';
+import qrcode from 'qrcode-terminal';
+import express from 'express';
+import { DialogKeywords } from '../../basic.gblib/services/DialogKeywords.js';
+import { GBAdminService } from '../../admin.gbapp/services/GBAdminService.js';
 
 /**
  * Support for Whatsapp.
  */
 export class WhatsappDirectLine extends GBService {
-
   public static conversationIds = {};
   public static mobiles = {};
   public static phones = {};
@@ -85,7 +79,7 @@ export class WhatsappDirectLine extends GBService {
   private browserWSEndpoint;
   private groupId;
 
-  constructor(
+  constructor (
     min: GBMinInstance,
     botId,
     directLineSecret,
@@ -102,24 +96,29 @@ export class WhatsappDirectLine extends GBService {
     this.whatsappServiceKey = whatsappServiceKey;
     this.whatsappServiceNumber = whatsappServiceNumber;
     this.whatsappServiceUrl = whatsappServiceUrl;
-    this.provider = whatsappServiceKey === "internal" ?
-      'GeneralBots' : whatsappServiceNumber.indexOf(';') > -1 ? 'maytapi' : 'chatapi';
+    this.provider =
+      whatsappServiceKey === 'internal'
+        ? 'GeneralBots'
+        : whatsappServiceNumber.indexOf(';') > -1
+        ? 'maytapi'
+        : 'chatapi';
     this.groupId = groupId;
   }
 
-  public static async asyncForEach(array, callback) {
+  public static async asyncForEach (array, callback) {
     for (let index = 0; index < array.length; index++) {
       await callback(array[index], index, array);
     }
   }
 
-  public async setup(setUrl) {
-
-    this.directLineClient =
-      new Swagger({
-        spec: JSON.parse(fs.readFileSync('directline-3.0.json', 'utf8')), usePromise: true
-      });
+  public async setup (setUrl) {
+    this.directLineClient = new Swagger({
+      spec: JSON.parse(Fs.readFileSync('directline-3.0.json', 'utf8')),
+      usePromise: true
+    });
     const client = await this.directLineClient;
+    let url;
+    let body;
 
     client.clientAuthorizations.add(
       'AuthorizationBotConnector',
@@ -129,25 +128,21 @@ export class WhatsappDirectLine extends GBService {
 
     switch (this.provider) {
       case 'GeneralBots':
-
         const minBoot = GBServer.globals.minBoot as any;
 
         // TODO:  REMOVE THIS.
         if (minBoot.botId !== this.botId) {
-
           this.customClient = minBoot.whatsAppDirectLine.customClient;
-
-        }
-        else {
-
+        } else {
           // Initialize the browser using a local profile for each bot.
 
           const gbaiName = `${this.min.botId}.gbai`;
           const localName = Path.join('work', gbaiName, 'profile');
 
-          const createClient = async (browserWSEndpoint) => {
+          const createClient = async browserWSEndpoint => {
             let puppeteer: any = {
-              headless: false, args: [
+              headless: false,
+              args: [
                 '--no-sandbox',
                 '--disable-setuid-sandbox',
                 '--disable-dev-shm-usage',
@@ -158,53 +153,58 @@ export class WhatsappDirectLine extends GBService {
                 '--disable-gpu',
                 '--disable-infobars',
                 '--disable-features=site-per-process',
-                `--user-data-dir=${localName}`]
+                `--user-data-dir=${localName}`
+              ]
             };
             if (browserWSEndpoint) {
               puppeteer = { browserWSEndpoint: browserWSEndpoint };
             }
 
-            const client = this.customClient = new Client({
-              authStrategy: new LocalAuth({
+            const client = (this.customClient = new wpp.Client({
+              authStrategy: new wpp.LocalAuth({
                 clientId: this.min.botId,
                 dataPath: localName
               }),
               puppeteer: puppeteer
-            });
+            }));
 
+            client.on(
+              'message',
+              (async message => {
+                await this.WhatsAppCallback(message, null);
+              }).bind(this)
+            );
 
-            client.on('message', (async message => {
-              await this.WhatsAppCallback(message, null);
-            }).bind(this));
+            client.on(
+              'qr',
+              (async qr => {
+                const adminNumber = this.min.core.getParam(this.min.instance, 'Bot Admin Number', null);
+                const adminEmail = this.min.core.getParam(this.min.instance, 'Bot Admin E-mail', null);
 
-            client.on('qr', (async (qr) => {
+                // Sends QR Code to boot bot admin.
 
-              const adminNumber = this.min.core.getParam(this.min.instance, 'Bot Admin Number', null);
-              const adminEmail = this.min.core.getParam(this.min.instance, 'Bot Admin E-mail', null);
+                const msg = `Please, scan QR Code with for bot ${this.botId}.`;
+                GBLog.info(msg);
+                qrcode.generate(qr, { small: true, scale: 0.5 });
 
-              // Sends QR Code to boot bot admin.
+                // While handling other bots uses boot instance of this class to send QR Codes.
 
-              const msg = `Please, scan QR Code with for bot ${this.botId}.`;
-              GBLog.info(msg);
-              qrcode.generate(qr, { small: true, scale: 0.5 });
+                const s = new DialogKeywords(this.min, null, null);
+                const qrBuf = await s.getQRCode(qr);
+                const gbaiName = `${this.min.botId}.gbai`;
+                const localName = Path.join('work', gbaiName, 'cache', `qr${GBAdminService.getRndReadableIdentifier()}.png`);
+                Fs.writeFileSync(localName, qrBuf);
+                const url = urlJoin(
+                  GBServer.globals.publicAddress,
+                  this.min.botId,
+                  'cache',
+                  Path.basename(localName)
+                );
+                GBServer.globals.minBoot.whatsAppDirectLine.sendFileToDevice(adminNumber, url, Path.basename(localName), msg);
 
-              // While handling other bots uses boot instance of this class to send QR Codes.
-
-              // const s = new DialogKeywords(min., null, null, null);
-              // const qrBuf = await s.getQRCode(qr);
-              // const gbaiName = `${this.min.botId}.gbai`;
-              // const localName = Path.join('work', gbaiName, 'cache', `qr${GBAdminService.getRndReadableIdentifier()}.png`);
-              // fs.writeFileSync(localName, qrBuf);
-              // const url = urlJoin(
-              //   GBServer.globals.publicAddress,
-              //   this.min.botId,
-              //   'cache',
-              //   Path.basename(localName)
-              // );
-              // GBServer.globals.minBoot.whatsAppDirectLine.sendFileToDevice(adminNumber, url, Path.basename(localName), msg);
-              // s.sendEmail(adminEmail, `Check your WhatsApp for bot ${this.botId}`, msg);
-
-            }).bind(this));
+                s.sendEmail({to:adminEmail, subject:`Check your WhatsApp for bot ${this.botId}`, body:msg);
+              }).bind(this)
+            );
 
             client.on('authenticated', async () => {
               this.browserWSEndpoint = client.pupBrowser.wsEndpoint();
@@ -212,11 +212,13 @@ export class WhatsappDirectLine extends GBService {
             });
 
             client.on('ready', async () => {
-
-              client.pupBrowser.on('disconnected', (async () => {
-                GBLog.info(`Browser terminated. Restarting ${this.min.botId} WhatsApp native provider.`);
-                await (createClient.bind(this))(null);
-              }).bind(this));
+              client.pupBrowser.on(
+                'disconnected',
+                (async () => {
+                  GBLog.info(`Browser terminated. Restarting ${this.min.botId} WhatsApp native provider.`);
+                  await createClient.bind(this)(null);
+                }).bind(this)
+              );
 
               GBLog.verbose(`GBWhatsApp: Emptying chat list for ${this.botId}...`);
 
@@ -224,7 +226,6 @@ export class WhatsappDirectLine extends GBService {
 
               const chats = await client.getChats();
               await CollectionUtil.asyncForEach(chats, async chat => {
-
                 const sleep = ms => {
                   return new Promise(resolve => {
                     setTimeout(resolve, ms);
@@ -234,27 +235,25 @@ export class WhatsappDirectLine extends GBService {
                 await sleep(wait);
                 if (chat.isGroup) {
                   await chat.clearMessages();
-                }
-                else if (!chat.pinned) {
+                } else if (!chat.pinned) {
                   await chat.delete();
                 }
               });
-
             });
 
             client.initialize();
           };
-          await (createClient.bind(this))(this.browserWSEndpoint);
+          await createClient.bind(this)(this.browserWSEndpoint);
 
           setUrl = false;
         }
         break;
 
       case 'chatapi':
-
+        url = urlJoin(this.whatsappServiceUrl, 'webhook')
         options = {
           method: 'POST',
-          url: urlJoin(this.whatsappServiceUrl, 'webhook'),
+          url: url,
           timeout: 10000,
           qs: {
             token: this.whatsappServiceKey,
@@ -268,35 +267,36 @@ export class WhatsappDirectLine extends GBService {
 
         break;
       case 'maytapi':
-
         let phoneId = this.whatsappServiceNumber.split(';')[0];
-        let productId = this.whatsappServiceNumber.split(';')[1]
-        let url = `${this.INSTANCE_URL}/${productId}/${phoneId}/config`;
+        let productId = this.whatsappServiceNumber.split(';')[1];
+        url = `${this.INSTANCE_URL}/${productId}/${phoneId}/config`;
+        body= {
+          webhook: `${GBServer.globals.publicAddress}/webhooks/whatsapp/${this.botId}`,
+          ack_delivery: false
+        };
         WhatsappDirectLine.phones[phoneId] = this.botId;
 
         options = {
           url: url,
           method: 'POST',
-          body: {
-            webhook: `${GBServer.globals.publicAddress}/webhooks/whatsapp/${this.botId}`,
-            "ack_delivery": false
-          },
+          body:body,
           headers: {
             'x-maytapi-key': this.whatsappServiceKey,
-            'Content-Type': 'application/json',
+            'Content-Type': 'application/json'
           },
-          json: true,
+          json: true
         };
         break;
     }
 
-    if (setUrl && options) {
-      const express = require('express');
+    if (setUrl && options && this.whatsappServiceUrl) {
       GBServer.globals.server.use(`/audios`, express.static('work'));
 
       if (options) {
         try {
-          await request.post(options);
+
+          const response: Response = await fetch(url,options);
+        
         } catch (error) {
           GBLog.error(`Error initializing 3rd party Whatsapp provider(1) ${error.message}`);
         }
@@ -304,11 +304,11 @@ export class WhatsappDirectLine extends GBService {
     }
   }
 
-  public async resetConversationId(botId, number, group = '') {
+  public async resetConversationId (botId, number, group = '') {
     WhatsappDirectLine.conversationIds[botId + number + group] = undefined;
   }
 
-  public async check() {
+  public async check () {
     switch (this.provider) {
       case 'GeneralBots':
         return true;
@@ -326,17 +326,15 @@ export class WhatsappDirectLine extends GBService {
     }
   }
 
-  public static providerFromRequest(req) {
-    return req.body.messages ? 'chatapi' :
-      req.body.message ? 'maytapi' : 'GeneralBots';
+  public static providerFromRequest (req) {
+    return req.body.messages ? 'chatapi' : req.body.message ? 'maytapi' : 'GeneralBots';
   }
 
-  public async received(req, res) {
-
+  public async received (req, res) {
     const provider = WhatsappDirectLine.providerFromRequest(req);
 
     let message, from, fromName, text;
-    let group = "";
+    let group = '';
     let answerText = null;
     let attachments = null;
 
@@ -344,19 +342,17 @@ export class WhatsappDirectLine extends GBService {
       case 'GeneralBots':
         message = req;
         text = message.body;
-        from = message.from.endsWith('@g.us') ?
-          message.author.split('@')[0] : message.from.split('@')[0];
+        from = message.from.endsWith('@g.us') ? message.author.split('@')[0] : message.from.split('@')[0];
         fromName = message._data.notifyName;
 
         if (message.hasMedia) {
           const base64Image = await message.downloadMedia();
           attachments = [];
-          attachments.push(
-            {
-              name: 'uploaded.png',
-              contentType: base64Image.mimetype,
-              contentUrl: `data:${base64Image.mimetype};base64,${base64Image.data}`
-            });
+          attachments.push({
+            name: 'uploaded.png',
+            contentType: base64Image.mimetype,
+            contentUrl: `data:${base64Image.mimetype};base64,${base64Image.data}`
+          });
         }
 
         break;
@@ -369,12 +365,11 @@ export class WhatsappDirectLine extends GBService {
 
         if (message.type !== 'chat') {
           attachments = [];
-          attachments.push(
-            {
-              name: 'uploaded',
-              contentType: 'application/octet-stream',
-              contentUrl: message.body
-            });
+          attachments.push({
+            name: 'uploaded',
+            contentType: 'application/octet-stream',
+            contentUrl: message.body
+          });
         }
 
         if (req.body.messages[0].fromMe) {
@@ -405,18 +400,16 @@ export class WhatsappDirectLine extends GBService {
     let botGroupID = WhatsappDirectLine.botGroups[this.min.botId];
     let botShortcuts = this.min.core.getParam<string>(this.min.instance, 'WhatsApp Group Shortcuts', null);
     if (!botShortcuts) {
-      botShortcuts = new Array()
-    }
-    else {
+      botShortcuts = new Array();
+    } else {
       botShortcuts = botShortcuts.split(' ');
     }
 
-    if (provider === "chatapi") {
+    if (provider === 'chatapi') {
       if (message.chatName.charAt(0) !== '+') {
         group = message.chatName;
       }
-    }
-    else if (provider === "GeneralBots") {
+    } else if (provider === 'GeneralBots') {
       if (message.from.endsWith('@g.us')) {
         group = message.from;
       }
@@ -428,7 +421,6 @@ export class WhatsappDirectLine extends GBService {
       // Bot name must be specified on config.
 
       if (botGroupID === group) {
-
         // Shortcut has been mentioned?
 
         let found = false;
@@ -440,20 +432,17 @@ export class WhatsappDirectLine extends GBService {
             }
           });
 
-
           // Verify if it is a group cache answer.
 
           const questions = this.min['groupCache'];
           if (questions && questions.length > 0) {
             questions.forEach(q => {
               if (q.content === e1 && !found) {
-                const answer = this.min.kbService['getAnswerById'](this.min.instance.instanceId,
-                  q.answerId);
+                const answer = this.min.kbService['getAnswerById'](this.min.instance.instanceId, q.answerId);
                 answerText = answer.content;
               }
             });
           }
-
 
           // Ignore group messages without the mention to Bot.
 
@@ -464,10 +453,8 @@ export class WhatsappDirectLine extends GBService {
               return;
             }
           }
-
         });
       }
-
     }
 
     const botId = this.min.instance.botId;
@@ -478,8 +465,7 @@ export class WhatsappDirectLine extends GBService {
       await state.promise(null, text);
 
       return; // Exit here.
-    };
-
+    }
 
     // Processes .gbapp message interception.
 
@@ -488,8 +474,7 @@ export class WhatsappDirectLine extends GBService {
     });
 
     const sec = new SecService();
-    const user = await sec.ensureUser(this.min.instance.instanceId, from,
-      fromName, '', 'whatsapp', fromName, null);
+    const user = await sec.ensureUser(this.min.instance.instanceId, from, fromName, '', 'whatsapp', fromName, null);
     const locale = user.locale ? user.locale : 'pt';
 
     if (answerText) {
@@ -499,7 +484,6 @@ export class WhatsappDirectLine extends GBService {
     }
 
     if (message.type === 'ptt') {
-
       if (process.env.AUDIO_DISABLED !== 'true') {
         const options = {
           url: provider ? message.body : message.text,
@@ -512,11 +496,11 @@ export class WhatsappDirectLine extends GBService {
         text = await GBConversationalService.getTextFromAudioBuffer(
           this.min.instance.speechKey,
           this.min.instance.cloudLocation,
-          buf, locale
+          buf,
+          locale
         );
       } else {
-        await this.sendToDevice(user.userSystemId,
-          `No momento estou apenas conseguindo ler mensagens de texto.`, null);
+        await this.sendToDevice(user.userSystemId, `No momento estou apenas conseguindo ler mensagens de texto.`, null);
       }
     }
 
@@ -527,14 +511,11 @@ export class WhatsappDirectLine extends GBService {
     // Check if this message is from a Human Agent itself.
 
     if (user.agentMode === 'self') {
-
       // Check if there is someone being handled by this Human Agent.
 
       const manualUser = await sec.getUserFromAgentSystemId(from);
       if (manualUser === null) {
-
         await sec.updateHumanAgent(from, this.min.instance.instanceId, null);
-
       } else {
         const agent = await sec.getUserFromSystemId(user.agentSystemId);
 
@@ -544,23 +525,38 @@ export class WhatsappDirectLine extends GBService {
           const message = await this.min.kbService.getAnswerTextByMediaName(this.min.instance.instanceId, filename);
 
           if (message === null) {
-            await this.sendToDeviceEx(user.userSystemId, `File ${filename} not found in any .gbkb published. Check the name or publish again the associated .gbkb.`,
-              locale, null);
+            await this.sendToDeviceEx(
+              user.userSystemId,
+              `File ${filename} not found in any .gbkb published. Check the name or publish again the associated .gbkb.`,
+              locale,
+              null
+            );
           } else {
             await this.min.conversationalService.sendMarkdownToMobile(this.min, null, user.userSystemId, message);
           }
         } else if (text === '/qt') {
           // TODO: Transfers only in pt-br for now.
-          await this.sendToDeviceEx(manualUser.userSystemId,
-            Messages[this.locale].notify_end_transfer(this.min.instance.botId), locale, null);
+          await this.sendToDeviceEx(
+            manualUser.userSystemId,
+            Messages[this.locale].notify_end_transfer(this.min.instance.botId),
+            locale,
+            null
+          );
 
-          if (user.agentSystemId.charAt(2) === ":") { // Agent is from Teams.
-            await this.min.conversationalService['sendOnConversation'](this.min, agent, Messages[this.locale].notify_end_transfer(this.min.instance.botId));
-          }
-          else {
-            await this.sendToDeviceEx(user.agentSystemId,
-              Messages[this.locale].notify_end_transfer(this.min.instance.botId), locale, null);
-
+          if (user.agentSystemId.charAt(2) === ':') {
+            // Agent is from Teams.
+            await this.min.conversationalService['sendOnConversation'](
+              this.min,
+              agent,
+              Messages[this.locale].notify_end_transfer(this.min.instance.botId)
+            );
+          } else {
+            await this.sendToDeviceEx(
+              user.agentSystemId,
+              Messages[this.locale].notify_end_transfer(this.min.instance.botId),
+              locale,
+              null
+            );
           }
           await sec.updateHumanAgent(manualUser.userSystemId, this.min.instance.instanceId, null);
           await sec.updateHumanAgent(user.agentSystemId, this.min.instance.instanceId, null);
@@ -569,48 +565,50 @@ export class WhatsappDirectLine extends GBService {
           await this.sendToDeviceEx(manualUser.userSystemId, `AGENTE: *${text}*`, locale, null);
         }
       }
-
-
     } else if (user.agentMode === 'human') {
-
       const agent = await sec.getUserFromSystemId(user.agentSystemId);
       if (text === '/t') {
-        await this.sendToDeviceEx(user.userSystemId, `Você já está sendo atendido por ${agent.userSystemId}.`, locale, null);
+        await this.sendToDeviceEx(
+          user.userSystemId,
+          `Você já está sendo atendido por ${agent.userSystemId}.`,
+          locale,
+          null
+        );
       } else if (text === '/qt' || text === 'Sair' || text === 'Fechar') {
         // TODO: Transfers only in pt-br for now.
         await this.endTransfer(from, locale, user, agent, sec);
       } else {
         GBLog.info(`USER (${from}) TO AGENT ${agent.userSystemId}: ${text}`);
 
-        if (user.agentSystemId.charAt(2) === ":" || agent.userSystemId.indexOf("@") > -1) { // Agent is from Teams or Google Chat.
+        if (user.agentSystemId.charAt(2) === ':' || agent.userSystemId.indexOf('@') > -1) {
+          // Agent is from Teams or Google Chat.
           await this.min.conversationalService['sendOnConversation'](this.min, agent, text);
+        } else {
+          await this.sendToDeviceEx(
+            user.agentSystemId,
+            `Bot: ${this.min.instance.botId}\n${from}: ${text}`,
+            locale,
+            null
+          );
         }
-        else {
-          await this.sendToDeviceEx(user.agentSystemId, `Bot: ${this.min.instance.botId}\n${from}: ${text}`, locale, null);
-        }
-
       }
-
     } else if (user.agentMode === 'bot' || user.agentMode === null || user.agentMode === undefined) {
-
       if (WhatsappDirectLine.conversationIds[botId + from + group] === undefined) {
         GBLog.info(`GBWhatsapp: Starting new conversation on Bot.`);
         const response = await client.Conversations.Conversations_StartConversation();
         const generatedConversationId = response.obj.conversationId;
 
         WhatsappDirectLine.conversationIds[botId + from + group] = generatedConversationId;
-        if (provider === "GeneralBots") {
+        if (provider === 'GeneralBots') {
           WhatsappDirectLine.chatIds[generatedConversationId] = message.from;
         }
         WhatsappDirectLine.mobiles[generatedConversationId] = from;
         WhatsappDirectLine.usernames[from] = fromName;
         WhatsappDirectLine.chatIds[generatedConversationId] = message.chatId;
 
-
         this.pollMessages(client, generatedConversationId, from, fromName);
         this.inputMessage(client, generatedConversationId, text, from, fromName, group, attachments);
       } else {
-
         this.inputMessage(client, conversationId, text, from, fromName, group, attachments);
       }
     } else {
@@ -622,21 +620,29 @@ export class WhatsappDirectLine extends GBService {
     }
   }
 
-  private async endTransfer(id: any, locale: string, user: GuaribasUser, agent: GuaribasUser, sec: SecService) {
-    await this.sendToDeviceEx(id,
-      Messages[this.locale].notify_end_transfer(this.min.instance.botId), locale, null);
+  private async endTransfer (id: any, locale: string, user: GuaribasUser, agent: GuaribasUser, sec: SecService) {
+    await this.sendToDeviceEx(id, Messages[this.locale].notify_end_transfer(this.min.instance.botId), locale, null);
 
-    if (user.agentSystemId.charAt(2) === ":") { // Agent is from Teams.
-      await this.min.conversationalService['sendOnConversation'](this.min, agent, Messages[this.locale].notify_end_transfer(this.min.instance.botId));
-    }
-    else {
-      await this.sendToDeviceEx(user.agentSystemId, Messages[this.locale].notify_end_transfer(this.min.instance.botId), locale, null);
+    if (user.agentSystemId.charAt(2) === ':') {
+      // Agent is from Teams.
+      await this.min.conversationalService['sendOnConversation'](
+        this.min,
+        agent,
+        Messages[this.locale].notify_end_transfer(this.min.instance.botId)
+      );
+    } else {
+      await this.sendToDeviceEx(
+        user.agentSystemId,
+        Messages[this.locale].notify_end_transfer(this.min.instance.botId),
+        locale,
+        null
+      );
     }
 
     await sec.updateHumanAgent(id, this.min.instance.instanceId, null);
   }
 
-  public inputMessage(client, conversationId, text, from, fromName, group, attachments) {
+  public inputMessage (client, conversationId, text, from, fromName, group, attachments) {
     return client.Conversations.Conversations_PostActivity({
       conversationId: conversationId,
       activity: {
@@ -655,7 +661,7 @@ export class WhatsappDirectLine extends GBService {
     });
   }
 
-  public pollMessages(client, conversationId, from, fromName) {
+  public pollMessages (client, conversationId, from, fromName) {
     GBLog.info(`GBWhatsapp: Starting message polling(${from}, ${conversationId}).`);
 
     let watermark: any;
@@ -669,14 +675,17 @@ export class WhatsappDirectLine extends GBService {
         watermark = response.obj.watermark;
         await this.printMessages(response.obj.activities, conversationId, from, fromName);
       } catch (err) {
-        GBLog.error(`Error calling printMessages on Whatsapp channel ${err.data === undefined ?
-          err : err.data} ${err.errObj ? err.errObj.message : ''}`);
+        GBLog.error(
+          `Error calling printMessages on Whatsapp channel ${err.data === undefined ? err : err.data} ${
+            err.errObj ? err.errObj.message : ''
+          }`
+        );
       }
     };
     setInterval(worker, this.pollInterval);
   }
 
-  public async printMessages(activities, conversationId, from, fromName) {
+  public async printMessages (activities, conversationId, from, fromName) {
     if (activities && activities.length) {
       // Ignore own messages.
 
@@ -692,7 +701,7 @@ export class WhatsappDirectLine extends GBService {
     }
   }
 
-  public async printMessage(activity, conversationId, from, fromName) {
+  public async printMessage (activity, conversationId, from, fromName) {
     let output = '';
 
     if (activity.text) {
@@ -720,21 +729,19 @@ export class WhatsappDirectLine extends GBService {
     await this.sendToDevice(from, output, conversationId);
   }
 
-  public renderHeroCard(attachment) {
+  public renderHeroCard (attachment) {
     return `${attachment.content.title} - ${attachment.content.text}`;
   }
 
-  public async sendFileToDevice(to, url, filename, caption, chatId) {
-
+  public async sendFileToDevice (to, url, filename, caption, chatId) {
     let options;
     switch (this.provider) {
       case 'GeneralBots':
-        const attachment = await MessageMedia.fromUrl(url);
+        const attachment = await wpp.MessageMedia.fromUrl(url);
         if (to.indexOf('@') == -1) {
           if (to.length == 18) {
             to = to + '@g.us';
-          }
-          else {
+          } else {
             to = to + '@c.us';
           }
         }
@@ -743,7 +750,6 @@ export class WhatsappDirectLine extends GBService {
         break;
 
       case 'chatapi':
-
         options = {
           method: 'POST',
           url: urlJoin(this.whatsappServiceUrl, 'sendFile'),
@@ -764,17 +770,16 @@ export class WhatsappDirectLine extends GBService {
 
         break;
       case 'maytapi':
-
         let contents = 0;
         let body = {
           to_number: to,
-          type: "media",
+          type: 'media',
           message: url,
           text: caption
         };
 
         let phoneId = this.whatsappServiceNumber.split(';')[0];
-        let productId = this.whatsappServiceNumber.split(';')[1]
+        let productId = this.whatsappServiceNumber.split(';')[1];
 
         options = {
           url: `${this.INSTANCE_URL}/${productId}/${phoneId}/sendMessage`,
@@ -783,8 +788,8 @@ export class WhatsappDirectLine extends GBService {
           body,
           headers: {
             'Content-Type': 'application/json',
-            'x-maytapi-key': this.whatsappServiceKey,
-          },
+            'x-maytapi-key': this.whatsappServiceKey
+          }
         };
 
         break;
@@ -800,19 +805,16 @@ export class WhatsappDirectLine extends GBService {
     }
   }
 
-  public async sendAudioToDevice(to, url, chatId) {
-
+  public async sendAudioToDevice (to, url, chatId) {
     let options;
     switch (this.provider) {
       case 'GeneralBots':
-
-        const attachment = MessageMedia.fromUrl(url);
+        const attachment = wpp.MessageMedia.fromUrl(url);
         await this.customClient.sendMessage(to, attachment);
 
         break;
 
       case 'chatapi':
-
         options = {
           method: 'POST',
           url: urlJoin(this.whatsappServiceUrl, 'sendPTT'),
@@ -829,7 +831,6 @@ export class WhatsappDirectLine extends GBService {
 
         break;
       case 'maytapi':
-
         options = {}; // TODO: Code Maytapi.
 
         break;
@@ -845,37 +846,30 @@ export class WhatsappDirectLine extends GBService {
     }
   }
 
-  public async sendTextAsAudioToDevice(to, msg, chatId) {
-
-    const url = await GBConversationalService.getAudioBufferFromText(
-      msg
-    );
+  public async sendTextAsAudioToDevice (to, msg, chatId) {
+    const url = await GBConversationalService.getAudioBufferFromText(msg);
 
     await this.sendFileToDevice(to, url, 'Audio', msg, chatId);
   }
 
-  public async sendToDevice(to: string, msg: string, conversationId) {
-
+  public async sendToDevice (to: string, msg: string, conversationId) {
     const cmd = '/audio ';
 
     let chatId = WhatsappDirectLine.chatIds[conversationId];
 
-    if (typeof (msg) !== 'object' && msg.startsWith(cmd)) {
+    if (typeof msg !== 'object' && msg.startsWith(cmd)) {
       msg = msg.substr(cmd.length);
 
       return await this.sendTextAsAudioToDevice(to, msg, chatId);
     } else {
-
       let options;
 
       switch (this.provider) {
         case 'GeneralBots':
-
           if (to.indexOf('@') == -1) {
             if (to.length == 18) {
               to = to + '@g.us';
-            }
-            else {
+            } else {
               to = to + '@c.us';
             }
           }
@@ -884,8 +878,6 @@ export class WhatsappDirectLine extends GBService {
           break;
 
         case 'chatapi':
-
-
           options = {
             method: 'POST',
             url: urlJoin(this.whatsappServiceUrl, 'message'),
@@ -902,7 +894,7 @@ export class WhatsappDirectLine extends GBService {
           break;
         case 'maytapi':
           let phoneId = this.whatsappServiceNumber.split(';')[0];
-          let productId = this.whatsappServiceNumber.split(';')[1]
+          let productId = this.whatsappServiceNumber.split(';')[1];
           let url = `${this.INSTANCE_URL}/${productId}/${phoneId}/sendMessage`;
 
           options = {
@@ -912,8 +904,8 @@ export class WhatsappDirectLine extends GBService {
             body: { type: 'text', message: msg, to_number: to },
             headers: {
               'Content-Type': 'application/json',
-              'x-maytapi-key': this.whatsappServiceKey,
-            },
+              'x-maytapi-key': this.whatsappServiceKey
+            }
           };
           break;
       }
@@ -922,8 +914,7 @@ export class WhatsappDirectLine extends GBService {
         try {
           GBLog.info(`Message [${msg}] is being sent to ${to}...`);
           await request.post(options);
-        }
-        catch (error) {
+        } catch (error) {
           GBLog.error(`Error sending message to Whatsapp provider ${error.message}`);
 
           // TODO: Handle Error: socket hang up and retry.
@@ -932,19 +923,13 @@ export class WhatsappDirectLine extends GBService {
     }
   }
 
-  public async sendToDeviceEx(to, text, locale, conversationId) {
-    text = await this.min.conversationalService.translate(
-      this.min,
-      text,
-      locale
-    );
+  public async sendToDeviceEx (to, text, locale, conversationId) {
+    text = await this.min.conversationalService.translate(this.min, text, locale);
     await this.sendToDevice(to, text, conversationId);
-
   }
 
-  private async WhatsAppCallback(req, res) {
+  private async WhatsAppCallback (req, res) {
     try {
-
       if (req.body && req.body.webhook) {
         res.status(200);
         res.end();
@@ -958,14 +943,11 @@ export class WhatsappDirectLine extends GBService {
       let botId;
       let text;
 
-
       switch (provider) {
-        case "GeneralBots":
-
+        case 'GeneralBots':
           // Ignore E2E messages used during initialization.
 
           if (req.type && req.type === 'e2e_notification') {
-
             return;
           }
 
@@ -973,11 +955,9 @@ export class WhatsappDirectLine extends GBService {
           senderName = req._data.notifyName;
           text = req.body;
 
-
           break;
 
-        case "chatapi":
-
+        case 'chatapi':
           if (req.body.ack) {
             res.status(200);
             res.end();
@@ -997,8 +977,7 @@ export class WhatsappDirectLine extends GBService {
             botId = GBConfigService.get('BOT_ID');
           }
           break;
-        case "maytapi":
-
+        case 'maytapi':
           if (req.body.type !== 'message') {
             res.status(200);
             res.end();
@@ -1023,8 +1002,7 @@ export class WhatsappDirectLine extends GBService {
       let user = await sec.getUserFromSystemId(id);
       GBLog.info(`A WhatsApp mobile requested instance for: ${botId}.`);
 
-      let urlMin: any = GBServer.globals.minInstances.filter
-        (p => p.instance.botId === botId)[0];
+      let urlMin: any = GBServer.globals.minInstances.filter(p => p.instance.botId === botId)[0];
 
       const botNumber = urlMin ? urlMin.core.getParam(urlMin.instance, 'Bot Number', null) : null;
       let activeMin;
@@ -1034,17 +1012,14 @@ export class WhatsappDirectLine extends GBService {
       text = text.replace(/\@\d+ /gi, '');
 
       let group;
-      if (provider === "chatapi") {
-
+      if (provider === 'chatapi') {
         // Ensures that the bot group is the active bot for the user (like switching).
 
         const message = req.body.messages[0];
         if (message.chatName.charAt(0) !== '+') {
           group = message.chatName;
         }
-      }
-      else if (provider === "GeneralBots") {
-
+      } else if (provider === 'GeneralBots') {
         // Ensures that the bot group is the active bot for the user (like switching).
 
         const message = req;
@@ -1055,17 +1030,15 @@ export class WhatsappDirectLine extends GBService {
 
       if (group) {
         GBLog.info(`Group: ${group}`);
-        function getKeyByValue(object, value) {
+        function getKeyByValue (object, value) {
           return Object.keys(object).find(key => object[key] === value);
         }
         const botId = getKeyByValue(WhatsappDirectLine.botGroups, group);
-        if (botId && user.instanceId !== this.min.instance.instanceId || !user) {
+        if ((botId && user.instanceId !== this.min.instance.instanceId) || !user) {
           user = await sec.ensureUser(this.min.instance.instanceId, id, senderName, '', 'whatsApp', senderName, null);
-
         }
         if (botId) {
-          activeMin = GBServer.globals.minInstances.filter
-            (p => p.instance.botId === botId)[0];
+          activeMin = GBServer.globals.minInstances.filter(p => p.instance.botId === botId)[0];
           await (activeMin as any).whatsAppDirectLine.received(req, res);
           return; // EXIT HERE.
         }
@@ -1074,7 +1047,6 @@ export class WhatsappDirectLine extends GBService {
       // Detects if the welcome message is enabled.
 
       if (process.env.WHATSAPP_WELCOME_DISABLED !== 'true') {
-
         // Tries to find if user wants to switch bots.
 
         let toSwitchMin = GBServer.globals.minInstances.filter(
@@ -1088,8 +1060,7 @@ export class WhatsappDirectLine extends GBService {
 
         // If bot has a fixed Find active bot instance.
 
-        activeMin = botNumber ? urlMin :
-          toSwitchMin ? toSwitchMin : GBServer.globals.minBoot;
+        activeMin = botNumber ? urlMin : toSwitchMin ? toSwitchMin : GBServer.globals.minBoot;
 
         // If it is the first time for the user, tries to auto-execute
         // start dialog if any is specified in Config.xlsx.
@@ -1097,19 +1068,17 @@ export class WhatsappDirectLine extends GBService {
         if (user === null || user.hearOnDialog) {
           user = await sec.ensureUser(activeMin.instance.instanceId, id, senderName, '', 'whatsapp', senderName, null);
 
-          const startDialog = user.hearOnDialog ?
-            user.hearOnDialog :
-            activeMin.core.getParam(activeMin.instance, 'Start Dialog', null);
+          const startDialog = user.hearOnDialog
+            ? user.hearOnDialog
+            : activeMin.core.getParam(activeMin.instance, 'Start Dialog', null);
 
           if (startDialog) {
             GBLog.info(`Calling /start to Auto start ${startDialog} for ${activeMin.instance.instanceId}...`);
-            if (provider === "chatapi") {
+            if (provider === 'chatapi') {
               req.body.messages[0].body = `/start`;
-            }
-            else if (provider === "maytapi") {
+            } else if (provider === 'maytapi') {
               req.body.message = `/start`;
-            }
-            else {
+            } else {
               req.body = `/start`;
             }
 
@@ -1121,19 +1090,19 @@ export class WhatsappDirectLine extends GBService {
           } else {
             await (activeMin as any).whatsAppDirectLine.sendToDevice(
               id,
-              `Olá! Seja bem-vinda(o)!\nMe chamo ${activeMin.instance.title}. Como posso ajudar? Pode me falar que eu te ouço, me manda um aúdio.`
-              , null);
+              `Olá! Seja bem-vinda(o)!\nMe chamo ${
+                activeMin.instance.title
+              }. Como posso ajudar? Pode me falar que eu te ouço, me manda um aúdio.`,
+              null
+            );
             if (res) {
               res.end();
             }
           }
-
         } else {
-
           // User wants to switch bots.
 
           if (toSwitchMin !== undefined) {
-
             // So gets the new bot instance information and prepares to
             // auto start dialog if any is specified.
 
@@ -1142,16 +1111,13 @@ export class WhatsappDirectLine extends GBService {
             await (activeMin as any).whatsAppDirectLine.resetConversationId(activeMin.botId, id, '');
             const startDialog = activeMin.core.getParam(activeMin.instance, 'Start Dialog', null);
 
-
             if (startDialog) {
               GBLog.info(`Calling /start for Auto start : ${startDialog} for ${activeMin.instance.botId}...`);
-              if (provider === "chatapi") {
+              if (provider === 'chatapi') {
                 req.body.messages[0].body = `/start`;
-              }
-              else if (provider === "maytapi") {
+              } else if (provider === 'maytapi') {
                 req.body.message = `/start`;
-              }
-              else {
+              } else {
                 req.body = `/start`;
               }
 
@@ -1166,14 +1132,15 @@ export class WhatsappDirectLine extends GBService {
             if (res) {
               res.end();
             }
-
           } else {
             activeMin = GBServer.globals.minInstances.filter(p => p.instance.instanceId === user.instanceId)[0];
             if (activeMin === undefined) {
               activeMin = GBServer.globals.minBoot;
               await (activeMin as any).whatsAppDirectLine.sendToDevice(
                 id,
-                `O outro Bot que você estava falando(${user.instanceId}), não está mais disponível. Agora você está falando comigo, ${activeMin.instance.title}...`
+                `O outro Bot que você estava falando(${
+                  user.instanceId
+                }), não está mais disponível. Agora você está falando comigo, ${activeMin.instance.title}...`
               );
             }
             await (activeMin as any).whatsAppDirectLine.received(req, res);
@@ -1184,13 +1151,11 @@ export class WhatsappDirectLine extends GBService {
           p => p.instance.botId.toLowerCase() === botId.toLowerCase()
         )[0];
 
-
         // Just pass the message to the receiver.
 
         await minInstance.whatsAppDirectLine.received(req, res);
       }
     } catch (error) {
-
       GBLog.error(`Error on Whatsapp callback: ${error.data ? error.data : error} ${error.stack}`);
     }
   }
