@@ -36,24 +36,24 @@
 
 'use strict';
 
-const express = require('express');
-const bodyParser = require('body-parser');
-const https = require('https');
-const mkdirp = require('mkdirp');
-const Path = require('path');
-
-import * as fs from 'fs';
+import express from 'express';
+import bodyParser from 'body-parser';
+import https from 'https';
+import mkdirp from 'mkdirp';
+import Path from 'path';
+import * as Fs from 'fs';
 import { GBLog, GBMinInstance, IGBCoreService, IGBInstance, IGBPackage } from 'botlib';
-import { GBAdminService } from '../packages/admin.gbapp/services/GBAdminService';
-import { AzureDeployerService } from '../packages/azuredeployer.gbapp/services/AzureDeployerService';
-import { GBConfigService } from '../packages/core.gbapp/services/GBConfigService';
-import { GBConversationalService } from '../packages/core.gbapp/services/GBConversationalService';
-import { GBCoreService } from '../packages/core.gbapp/services/GBCoreService';
-import { GBDeployer } from '../packages/core.gbapp/services/GBDeployer';
-import { GBImporter } from '../packages/core.gbapp/services/GBImporterService';
-import { GBMinService } from '../packages/core.gbapp/services/GBMinService';
-var auth = require('basic-auth');
-
+import { GBAdminService } from '../packages/admin.gbapp/services/GBAdminService.js';
+import { AzureDeployerService } from '../packages/azuredeployer.gbapp/services/AzureDeployerService.js';
+import { GBConfigService } from '../packages/core.gbapp/services/GBConfigService.js';
+import { GBConversationalService } from '../packages/core.gbapp/services/GBConversationalService.js';
+import { GBCoreService } from '../packages/core.gbapp/services/GBCoreService.js';
+import { GBDeployer } from '../packages/core.gbapp/services/GBDeployer.js';
+import { GBImporter } from '../packages/core.gbapp/services/GBImporterService.js';
+import { GBMinService } from '../packages/core.gbapp/services/GBMinService.js';
+import auth from 'basic-auth';
+import child_process from 'child_process';
+import * as winston from 'winston-logs-display';
 
 /**
  * Global shared server data;
@@ -70,6 +70,7 @@ export class RootData {
   public wwwroot: string; // .gbui or a static webapp.
   public entryPointDialog: string; // To replace default welcome dialog.
   public debugConversationId: any; // Used to self-message during debug.
+  public debuggers: any[]; // Client of attached Debugger instances by botId.
 }
 /**
  * General Bots open-core entry point.
@@ -81,37 +82,44 @@ export class GBServer {
    *  Program entry-point.
    */
 
-  public static run() {
-    
-
+  public static run () {
     GBLog.info(`The Bot Server is in STARTING mode...`);
     GBServer.globals = new RootData();
     GBConfigService.init();
     const port = GBConfigService.getServerPort();
+
+    if (process.env.TEST_SHELL) {
+      GBLog.info(`Running TEST_SHELL: ${process.env.TEST_SHELL}...`);
+      try {
+        child_process.execSync(process.env.TEST_SHELL);
+      } catch (error) {
+        GBLog.error(`Running TEST_SHELL ERROR: ${error}...`);
+      }
+    }
+
     const server = express();
+
     GBServer.globals.server = server;
     GBServer.globals.appPackages = [];
     GBServer.globals.sysPackages = [];
     GBServer.globals.minInstances = [];
     GBServer.globals.wwwroot = null;
     GBServer.globals.entryPointDialog = null;
+    GBServer.globals.debuggers = [];
 
     server.use(bodyParser.json());
     server.use(bodyParser.urlencoded({ extended: true }));
 
-
     // Creates working directory.
 
     const workDir = Path.join(process.env.PWD, 'work');
-    if (!fs.existsSync(workDir)) {
+    if (!Fs.existsSync(workDir)) {
       mkdirp.sync(workDir);
     }
 
     const mainCallback = () => {
       (async () => {
-
         try {
-
           GBLog.info(`Now accepting connections on ${port}...`);
           process.env['NODE_TLS_REJECT_UNAUTHORIZED'] = '0';
 
@@ -146,9 +154,11 @@ export class GBServer {
             await core.initStorage();
           } catch (error) {
             GBLog.verbose(`Error initializing storage: ${error}`);
-            GBServer.globals.bootInstance =
-              await core.createBootInstance(core, azureDeployer, GBServer.globals.publicAddress);
-
+            GBServer.globals.bootInstance = await core.createBootInstance(
+              core,
+              azureDeployer,
+              GBServer.globals.publicAddress
+            );
           }
 
           core.ensureAdminIsSecured();
@@ -176,7 +186,6 @@ export class GBServer {
           );
 
           if (instances.length === 0) {
-
             const instance = await importer.importIfNotExistsBotPackage(
               GBConfigService.get('BOT_ID'),
               'boot.gbot',
@@ -201,32 +210,32 @@ export class GBServer {
           await minService.buildMin(instances);
 
           if (process.env.ENABLE_WEBLOG) {
-            var admins = {
-              'admin': { password: process.env.ADMIN_PASS },
+            const admins = {
+              admin: { password: process.env.ADMIN_PASS }
             };
 
             // ... some not authenticated middlewares
 
-            server.use((req, res, next) => {
+            server.use(async (req, res, next) => {
               if (req.originalUrl.startsWith('/logs')) {
-                var user = auth(req);
+                const user = auth(req);
                 if (!user || !admins[user.name] || admins[user.name].password !== user.pass) {
                   res.set('WWW-Authenticate', 'Basic realm="example"');
                   return res.status(401).send();
                 }
+              } else {
+                return next();
               }
-              return next();
             });
 
             // If global log enabled, reorders transports adding web logging.
 
             const loggers = GBLog.getLogger();
-            require('winston-logs-display')(server, loggers[1]);
+            winston.default(server, loggers[1]);
           }
 
-
           GBLog.info(`The Bot Server is in RUNNING mode...`);
-          
+
           // Opens Navigator.
 
           // TODO: Config: core.openBrowserInDevelopment();
@@ -237,22 +246,20 @@ export class GBServer {
       })();
     };
     if (process.env.CERTIFICATE_PFX) {
-      let options = {
+      const options = {
         passphrase: process.env.CERTIFICATE_PASSPHRASE,
-        pfx: fs.readFileSync(process.env.CERTIFICATE_PFX)
+        pfx: Fs.readFileSync(process.env.CERTIFICATE_PFX)
       };
       const httpsServer = https.createServer(options, server).listen(port, mainCallback);
 
       if (process.env.CERTIFICATE2_PFX) {
-        let options = {
+        const options = {
           passphrase: process.env.CERTIFICATE2_PASSPHRASE,
-          pfx: fs.readFileSync(process.env.CERTIFICATE2_PFX)
+          pfx: Fs.readFileSync(process.env.CERTIFICATE2_PFX)
         };
         httpsServer.addContext(process.env.CERTIFICATE2_DOMAIN, options);
-
       }
-    }
-    else {
+    } else {
       server.listen(port, mainCallback);
     }
   }
