@@ -38,8 +38,8 @@
 
 import { GBDialogStep, GBLog, GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
 import { Sequelize } from 'sequelize-typescript';
-import { AdminDialog } from './dialogs/AdminDialog';
-import { GuaribasAdmin } from './models/AdminModel';
+import { AdminDialog } from './dialogs/AdminDialog.js';
+import { GuaribasAdmin } from './models/AdminModel.js';
 
 /**
  * The package for admin.gbapp.
@@ -47,28 +47,27 @@ import { GuaribasAdmin } from './models/AdminModel';
 export class GBAdminPackage implements IGBPackage {
   public sysPackages: IGBPackage[];
 
-  public async getDialogs(min: GBMinInstance) {
+  public async getDialogs (min: GBMinInstance) {
     GBLog.verbose(`getDialogs called.`);
   }
-  public async unloadPackage(core: IGBCoreService): Promise<void> {
+  public async unloadPackage (core: IGBCoreService): Promise<void> {
     GBLog.verbose(`unloadPackage called.`);
   }
-  public async unloadBot(min: GBMinInstance): Promise<void> {
+  public async unloadBot (min: GBMinInstance): Promise<void> {
     GBLog.verbose(`unloadBot called.`);
   }
-  public async onNewSession(min: GBMinInstance, step: GBDialogStep): Promise<void> {
+  public async onNewSession (min: GBMinInstance, step: GBDialogStep): Promise<void> {
     GBLog.verbose(`onNewSession called.`);
   }
-  public async onExchangeData(min: GBMinInstance, kind: string, data: any) {
+  public async onExchangeData (min: GBMinInstance, kind: string, data: any) {
     GBLog.verbose(`onExchangeData called.`);
   }
 
-  public async loadPackage(core: IGBCoreService, sequelize: Sequelize): Promise<void> {
+  public async loadPackage (core: IGBCoreService, sequelize: Sequelize): Promise<void> {
     core.sequelize.addModels([GuaribasAdmin]);
   }
 
-  public async loadBot(min: GBMinInstance): Promise<void> {
+  public async loadBot (min: GBMinInstance): Promise<void> {
     AdminDialog.setup(min);
   }
-
 }
