@@ -3,7 +3,7 @@ export const Messages = {
     authenticate: 'Please, authenticate:',
     welcome: 'Welcome to Pragmatismo.io GeneralBots Administration.',
     which_task: 'Which task do you wanna run now?',
-    working: (command) => `I'm working on ${command}...`,
+    working: command => `I'm working on ${command}...`,
     finished_working: 'Done.',
     unknown_command: text =>
       `Well, but ${text} is not a administrative General Bots command, I will try to search for it.`,
@@ -12,7 +12,7 @@ export const Messages = {
     deployPackage: text => `Deploying package ${text}...`,
     redeployPackage: text => `Redeploying package ${text}...`,
     packageUndeployed: text => `√ Package ${text} undeployed...`,
-    consent: (url) => `Please, consent access to this app at: [Microsoft Online](${url}).`,
+    consent: url => `Please, consent access to this app at: [Microsoft Online](${url}).`,
     wrong_password: 'Sorry, wrong password. Please, try again.',
     enter_authenticator_tenant: 'Enter the Authenticator Tenant (eg.: domain.onmicrosoft.com):',
     enter_authenticator_authority_host_url: 'Enter the Authority Host URL (eg.: https://login.microsoftonline.com): ',
@@ -28,7 +28,7 @@ export const Messages = {
     authenticate: 'Please, authenticate:',
     welcome: 'Welcome to Pragmatismo.io GeneralBots Administration.',
     which_task: 'Which task do you wanna run now?',
-    working: (command) => `I'm working on ${command}...`,
+    working: command => `I'm working on ${command}...`,
     finished_working: 'Done.',
     unknown_command: text =>
       `Well, but ${text} is not a administrative General Bots command, I will try to search for it.`,
@@ -37,7 +37,7 @@ export const Messages = {
     deployPackage: text => `Deploying package ${text}...`,
     redeployPackage: text => `Redeploying package ${text}...`,
     packageUndeployed: text => `Package ${text} undeployed...`,
-    consent: (url) => `Please, consent access to this app at: [Microsoft Online](${url}).`,
+    consent: url => `Please, consent access to this app at: [Microsoft Online](${url}).`,
     wrong_password: 'Sorry, wrong password. Please, try again.',
     enter_authenticator_tenant: 'Enter the Authenticator Tenant (eg.: domain.onmicrosoft.com):',
     enter_authenticator_authority_host_url: 'Enter the Authority Host URL (eg.: https://login.microsoftonline.com): ',
