@@ -38,35 +38,33 @@
 
 import { GBDialogStep, GBLog, GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
 import { Sequelize } from 'sequelize-typescript';
-import { GuaribasConversation, GuaribasConversationMessage } from './models';
+import { GuaribasConversation, GuaribasConversationMessage } from './models/index.js';
 
 /**
  * .gblib Package handler.
  */
 export class GBAnalyticsPackage implements IGBPackage {
   public sysPackages: IGBPackage[];
-  public async getDialogs(min: GBMinInstance) {
+  public async getDialogs (min: GBMinInstance) {
     GBLog.verbose(`getDialogs called.`);
   }
-  public async loadPackage(core: IGBCoreService, sequelize: Sequelize): Promise<void> {
+  public async loadPackage (core: IGBCoreService, sequelize: Sequelize): Promise<void> {
     GBLog.verbose(`loadPackage called.`);
     core.sequelize.addModels([GuaribasConversation, GuaribasConversationMessage]);
-
   }
-  public async unloadPackage(core: IGBCoreService): Promise<void> {
+  public async unloadPackage (core: IGBCoreService): Promise<void> {
     GBLog.verbose(`unloadPackage called.`);
   }
-  public async loadBot(min: GBMinInstance): Promise<void> {
+  public async loadBot (min: GBMinInstance): Promise<void> {
     GBLog.verbose(`loadBot called.`);
   }
-  public async unloadBot(min: GBMinInstance): Promise<void> {
+  public async unloadBot (min: GBMinInstance): Promise<void> {
     GBLog.verbose(`unloadBot called.`);
   }
-  public async onNewSession(min: GBMinInstance, step: GBDialogStep): Promise<void> {
+  public async onNewSession (min: GBMinInstance, step: GBDialogStep): Promise<void> {
     GBLog.verbose(`onNewSession called.`);
   }
-  public async onExchangeData(min: GBMinInstance, kind: string, data: any) {
+  public async onExchangeData (min: GBMinInstance, kind: string, data: any) {
     GBLog.verbose(`onExchangeData called.`);
   }
-
 }
