@@ -1,6 +1,6 @@
 export const Messages = {
   'en-US': {
-    about_suggestions:  'Suggestions are welcomed and improve my quality...'
+    about_suggestions: 'Suggestions are welcomed and improve my quality...'
   },
   'pt-BR': {
     about_suggestions: 'Sugestões melhoram muito minha qualidade...'
