@@ -33,6 +33,7 @@
 'use strict';
 
 import { GBLog } from 'botlib';
+import * as en from 'dotenv-extended';
 
 /**
  * @fileoverview General Bots server core.
@@ -42,10 +43,10 @@ import { GBLog } from 'botlib';
  * Base configuration for the server like storage.
  */
 export class GBConfigService {
-  public static getBoolean(value: string): boolean {
-    return this.get(value) as unknown as boolean;
+  public static getBoolean (value: string): boolean {
+    return (this.get(value) as unknown) as boolean;
   }
-  public static getServerPort(): string {
+  public static getServerPort (): string {
     if (process.env.PORT) {
       return process.env.PORT;
     }
@@ -56,9 +57,9 @@ export class GBConfigService {
     return '4242';
   }
 
-  public static init(): any {
+  public static init (): any {
     try {
-      require('dotenv-extended').load({
+      en.load({
         encoding: 'utf8',
         silent: true,
         path: '.env',
@@ -77,7 +78,7 @@ export class GBConfigService {
     }
   }
 
-  public static get(key: string): string | undefined {
+  public static get (key: string): string | undefined {
     let value = GBConfigService.tryGet(key);
 
     if (value === undefined) {
@@ -163,13 +164,12 @@ export class GBConfigService {
     return value;
   }
 
-  public static tryGet(key: string): any {
+  public static tryGet (key: string): any {
     let value = process.env[`container:${key}`];
     if (value === undefined) {
       value = process.env[key];
     }
 
-    
     return value;
   }
 }
