@@ -40,10 +40,10 @@ import { GBMinInstance, IGBDialog } from 'botlib';
 
 import { BotAdapter } from 'botbuilder';
 import { WaterfallDialog } from 'botbuilder-dialogs';
-import { AnalyticsService } from '../../analytics.gblib/services/AnalyticsService';
-import { GBConversationalService } from '../../core.gbapp/services/GBConversationalService';
-import { CSService } from '../services/CSService';
-import { Messages } from '../strings';
+import { AnalyticsService } from '../../analytics.gblib/services/AnalyticsService.js';
+import { GBConversationalService } from '../../core.gbapp/services/GBConversationalService.js';
+import { CSService } from '../services/CSService.js';
+import { Messages } from '../strings.js';
 
 /**
  * Dialog for collecting quality of answer.
@@ -55,60 +55,63 @@ export class QualityDialog extends IGBDialog {
    * @param bot The bot adapter.
    * @param min The minimal bot instance data.
    */
-  public static setup(bot: BotAdapter, min: GBMinInstance) {
+  public static setup (bot: BotAdapter, min: GBMinInstance) {
     const service = new CSService();
 
-    min.dialogs.add(new WaterfallDialog('/check', [
-      async step => {
-        const locale = step.context.activity.locale;
-        await min.conversationalService.sendText(min, step, Messages[locale].check_whatsapp_ok);
-        return await step.replaceDialog('/ask', { isReturning: true });
-      }
-    ]
-    ));
-
-    min.dialogs.add(new WaterfallDialog('/quality', [
-      async step => {
-        const locale = step.context.activity.locale;
-        const user = await min.userProfile.get(step.context, {});
-
-        const score = step.result;
-
-        if (score === 0) {
-          await min.conversationalService.sendText(min, step, Messages[locale].im_sorry_lets_try);
-
-          return await step.next();
-        } else {
-          await min.conversationalService.sendText(min, step, Messages[locale].great_thanks);
-          await min.conversationalService.sendEvent(min, step, 'play', {
-            playerType: 'markdown',
-            data: {
-              content: Messages[locale].great_thanks,
-            }
-          });
-          let sleep = ms => {
-            return new Promise(resolve => {
-              setTimeout(resolve, ms);
-            });
-          };
-
-          await service.insertQuestionAlternate(
-            min.instance.instanceId,
-            user.lastQuestion,
-            user.lastQuestionId
-          );
-
-          // Updates values to perform Bot Analytics.
-
-          const analytics = new AnalyticsService();
-          analytics.updateConversationSuggestion(
-            min.instance.instanceId, user.conversation, step.result, user.systemUser.locale);
-
-          // Goes to the ask loop.
-
-          return await step.replaceDialog('/ask', { emptyPrompt: true });
+    min.dialogs.add(
+      new WaterfallDialog('/check', [
+        async step => {
+          const locale = step.context.activity.locale;
+          await min.conversationalService.sendText(min, step, Messages[locale].check_whatsapp_ok);
+          return await step.replaceDialog('/ask', { isReturning: true });
         }
-      }
-    ]));
+      ])
+    );
+
+    min.dialogs.add(
+      new WaterfallDialog('/quality', [
+        async step => {
+          const locale = step.context.activity.locale;
+          const user = await min.userProfile.get(step.context, {});
+
+          const score = step.result;
+
+          if (score === 0) {
+            await min.conversationalService.sendText(min, step, Messages[locale].im_sorry_lets_try);
+
+            return await step.next();
+          } else {
+            await min.conversationalService.sendText(min, step, Messages[locale].great_thanks);
+            await min.conversationalService.sendEvent(min, step, 'play', {
+              playerType: 'markdown',
+              data: {
+                content: Messages[locale].great_thanks
+              }
+            });
+            let sleep = ms => {
+              return new Promise(resolve => {
+                setTimeout(resolve, ms);
+              });
+            };
+
+            await service.insertQuestionAlternate(min.instance.instanceId, user.lastQuestion, user.lastQuestionId);
+
+            // Updates values to perform Bot Analytics.
+
+            const analytics = new AnalyticsService();
+            analytics.updateConversationSuggestion(
+              min.instance.instanceId,
+              user.conversation,
+              step.result,
+              user.systemUser.locale
+            );
+
+            // Goes to the ask loop.
+
+            return await step.replaceDialog('/ask', { emptyPrompt: true });
+          }
+        }
+      ])
+    );
   }
 }
