@@ -1,8 +1,8 @@
 export const Messages = {
   'en-US': {
-    notify_end_transfer: (botName) => `Now talking to ${botName} again.`
+    notify_end_transfer: botName => `Now talking to ${botName} again.`
   },
   'pt-BR': {
-    notify_end_transfer: (botName) => `Falando com o bot ${botName} novamente.`
+    notify_end_transfer: botName => `Falando com o bot ${botName} novamente.`
   }
 };
