@@ -36,19 +36,19 @@
 
 'use strict';
 
-import { GBServer } from '../../../src/app';
+import { GBServer } from '../../../src/app.js';
 import { BotAdapter } from 'botbuilder';
 import { WaterfallDialog } from 'botbuilder-dialogs';
 import { GBLog, GBMinInstance, IGBDialog, IGBPackage } from 'botlib';
-import { Messages } from '../strings';
-import { KBService } from './../services/KBService';
-import { GuaribasAnswer } from '../models';
-import { SecService } from '../../security.gbapp/services/SecService';
+import { Messages } from '../strings.js';
+import { KBService } from './../services/KBService.js';
+import { GuaribasAnswer } from '../models/index.js';
+import { SecService } from '../../security.gbapp/services/SecService.js';
 import { CollectionUtil, AzureText } from 'pragmatismo-io-framework';
-import { GBVMService } from '../../basic.gblib/services/GBVMService';
-import { GBImporter } from '../../core.gbapp/services/GBImporterService';
-import { GBDeployer } from '../../core.gbapp/services/GBDeployer';
-import { GBConfigService } from '../../core.gbapp/services/GBConfigService';
+import { GBVMService } from '../../basic.gblib/services/GBVMService.js';
+import { GBImporter } from '../../core.gbapp/services/GBImporterService.js';
+import { GBDeployer } from '../../core.gbapp/services/GBDeployer.js';
+import { GBConfigService } from '../../core.gbapp/services/GBConfigService.js';
 
 /**
  * Dialog arguments.
@@ -69,7 +69,7 @@ export class AskDialog extends IGBDialog {
    * @param bot The bot adapter.
    * @param min The minimal bot instance data.
    */
-  public static setup(bot: BotAdapter, min: GBMinInstance) {
+  public static setup (bot: BotAdapter, min: GBMinInstance) {
     const service = new KBService(min.core.sequelize);
     const importer = new GBImporter(min.core);
     this.deployer = new GBDeployer(min.core, importer);
@@ -79,13 +79,12 @@ export class AskDialog extends IGBDialog {
     min.dialogs.add(new WaterfallDialog('/ask', AskDialog.getAskDialog(min)));
   }
 
-  private static getAskDialog(min: GBMinInstance) {
+  private static getAskDialog (min: GBMinInstance) {
     return [
       async step => {
         if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
           return await step.beginDialog('/auth');
-        }
-        else {
+        } else {
           return await step.next(step.options);
         }
       },
@@ -101,9 +100,9 @@ export class AskDialog extends IGBDialog {
         if (step.options && step.options.firstTime) {
           text = Messages[locale].ask_first_time;
         } else if (step.options && step.options.isReturning) {
-          text = ""; // REMOVED: Messages[locale].anything_else;
+          text = ''; // REMOVED: Messages[locale].anything_else;
         } else if (step.options && step.options.emptyPrompt) {
-          text = "";
+          text = '';
         } else if (user.subjects.length > 0) {
           text = Messages[locale].which_question;
         } else {
@@ -111,7 +110,6 @@ export class AskDialog extends IGBDialog {
         }
 
         return await min.conversationalService.prompt(min, step, text);
-
       },
       async step => {
         if (step.result) {
@@ -120,7 +118,15 @@ export class AskDialog extends IGBDialog {
 
           let sec = new SecService();
           const member = step.context.activity.from;
-          const user = await sec.ensureUser(min.instance.instanceId, member.id, member.name, '', 'web', member.name, null);
+          const user = await sec.ensureUser(
+            min.instance.instanceId,
+            member.id,
+            member.name,
+            '',
+            'web',
+            member.name,
+            null
+          );
 
           let handled = false;
           let nextDialog = null;
@@ -132,9 +138,7 @@ export class AskDialog extends IGBDialog {
             user: user ? user['dataValues'] : null
           };
           await CollectionUtil.asyncForEach(min.appPackages, async (e: IGBPackage) => {
-            if (
-              nextDialog = await e.onExchangeData(min, 'handleAnswer', data)
-            ) {
+            if ((nextDialog = await e.onExchangeData(min, 'handleAnswer', data))) {
               handled = true;
             }
           });
@@ -153,13 +157,12 @@ export class AskDialog extends IGBDialog {
     ];
   }
 
-  private static getAnswerDialog(min: GBMinInstance, service: KBService) {
+  private static getAnswerDialog (min: GBMinInstance, service: KBService) {
     return [
       async step => {
         if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
           return await step.beginDialog('/auth');
-        }
-        else {
+        } else {
           return await step.next(step.options);
         }
       },
@@ -176,10 +179,9 @@ export class AskDialog extends IGBDialog {
         // when people type just the @botName in MSTEAMS for example.
 
         if (!text) {
-          const startDialog =
-            min.core.getParam(min.instance, 'Start Dialog', null);
+          const startDialog = min.core.getParam(min.instance, 'Start Dialog', null);
           if (startDialog) {
-            await GBVMService.callVM(startDialog.toLowerCase().trim(), min, step, this.deployer);
+            await GBVMService.callVM(startDialog.toLowerCase().trim(), min, step, this.deployer, false);
           }
 
           return step.endDialog();
@@ -189,7 +191,7 @@ export class AskDialog extends IGBDialog {
 
         // Stops any content on projector.
         if (step.context.activity.channelId !== 'msteams') {
-            await min.conversationalService.sendEvent(min, step, 'stop', undefined);
+          await min.conversationalService.sendEvent(min, step, 'stop', undefined);
         }
         // Handle extra text from FAQ.
 
@@ -201,8 +203,11 @@ export class AskDialog extends IGBDialog {
 
         // Searches KB for the first time.
 
-        const searchScore = min.core.getParam(min.instance, 'Search Score',
-          min.instance.searchScore ? min.instance.searchScore : minBoot.instance.searchScore);
+        const searchScore = min.core.getParam(
+          min.instance,
+          'Search Score',
+          min.instance.searchScore ? min.instance.searchScore : minBoot.instance.searchScore
+        );
 
         user.lastQuestion = text;
         await min.userProfile.set(step.context, user);
@@ -256,7 +261,7 @@ export class AskDialog extends IGBDialog {
 
         // Tries to answer by NLP.
 
-        let nextDialog = await min.conversationalService["routeNLP2"](step, min, text);
+        let nextDialog = await min.conversationalService['routeNLP2'](step, min, text);
         if (nextDialog) {
           return nextDialog;
         }
@@ -270,58 +275,54 @@ export class AskDialog extends IGBDialog {
           const docs = await min.kbService['getDocs'](min.instance.instanceId);
 
           await CollectionUtil.asyncForEach(docs, async (doc: GuaribasAnswer) => {
-
             if (!answered) {
-
               const answerText = await min.kbService['readComprehension'](min.instance.instanceId, doc.content, text);
               answered = true;
-              text = await min.conversationalService.translate(min, text, user.systemUser.locale
-                ? user.systemUser.locale
-                : min.core.getParam<string>(min.instance, 'Locale', GBConfigService.get('LOCALE')));
+              text = await min.conversationalService.translate(
+                min,
+                text,
+                user.systemUser.locale
+                  ? user.systemUser.locale
+                  : min.core.getParam<string>(min.instance, 'Locale', GBConfigService.get('LOCALE'))
+              );
               await min.conversationalService.sendText(min, step, answerText);
               await min.conversationalService.sendEvent(min, step, 'stop', undefined);
             }
-
           });
           return await step.replaceDialog('/ask', { isReturning: true });
         }
 
         // Not found.
 
-        const message = min.core.getParam<string>(min.instance, 'Not Found Message',
-          Messages[locale].did_not_find);
+        const message = min.core.getParam<string>(min.instance, 'Not Found Message', Messages[locale].did_not_find);
 
         await min.conversationalService.sendText(min, step, message);
         return await step.replaceDialog('/ask', { isReturning: true });
-
       }
     ];
   }
 
-  private static async handleAnswer(service: KBService, min: GBMinInstance, step: any, answer: GuaribasAnswer) {
+  private static async handleAnswer (service: KBService, min: GBMinInstance, step: any, answer: GuaribasAnswer) {
     const text = answer.content;
     if (text.endsWith('.docx')) {
-
-
       const mainName = GBVMService.getMethodNameFromVBSFilename(text);
-      return await GBVMService.callVM(mainName, min, step, this.deployer);
+      return await GBVMService.callVM(mainName, min, step, this.deployer, false);
     } else {
       await service.sendAnswer(min, AskDialog.getChannel(step), step, answer);
       return await step.replaceDialog('/ask', { isReturning: true });
     }
   }
 
-  private static getChannel(step): string {
+  private static getChannel (step): string {
     return !isNaN(step.context.activity['mobile']) ? 'whatsapp' : step.context.activity.channelId;
   }
 
-  private static getAnswerEventDialog(service: KBService, min: GBMinInstance) {
+  private static getAnswerEventDialog (service: KBService, min: GBMinInstance) {
     return [
       async step => {
         if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
           return await step.beginDialog('/auth');
-        }
-        else {
+        } else {
           return await step.next(step.options);
         }
       },
@@ -330,7 +331,7 @@ export class AskDialog extends IGBDialog {
         const data = step.options as AskDialogArgs;
         if (data !== undefined && data.questionId !== undefined) {
           const question = await service.getQuestionById(min.instance.instanceId, data.questionId);
-          const answer = await service.getAnswerById(min.instance.instanceId, question.answerId );
+          const answer = await service.getAnswerById(min.instance.instanceId, question.answerId);
           // Sends the answer to all outputs, including projector.
           await service.sendAnswer(min, AskDialog.getChannel(step), step, answer);
           await step.replaceDialog('/ask', { isReturning: true });
