@@ -39,9 +39,9 @@
 import { BotAdapter } from 'botbuilder';
 import { WaterfallDialog } from 'botbuilder-dialogs';
 import { GBMinInstance, IGBDialog } from 'botlib';
-import { GBConversationalService } from '../../core.gbapp/services/GBConversationalService';
-import { Messages } from '../strings';
-import { KBService } from './../services/KBService';
+import { GBConversationalService } from '../../core.gbapp/services/GBConversationalService.js';
+import { Messages } from '../strings.js';
+import { KBService } from './../services/KBService.js';
 
 /**
  * Handle display of FAQ allowing direct access to KB.
@@ -53,34 +53,34 @@ export class FaqDialog extends IGBDialog {
    * @param bot The bot adapter.
    * @param min The minimal bot instance data.
    */
-  public static setup(bot: BotAdapter, min: GBMinInstance) {
-
+  public static setup (bot: BotAdapter, min: GBMinInstance) {
     const service = new KBService(min.core.sequelize);
 
-    min.dialogs.add(new WaterfallDialog('/faq', [
-      async step => {
-        if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
-          return await step.beginDialog('/auth');
-        }
-        else{
-          return await step.next(step.options);
-        }
-      },
+    min.dialogs.add(
+      new WaterfallDialog('/faq', [
+        async step => {
+          if (step.context.activity.channelId !== 'msteams' && process.env.ENABLE_AUTH) {
+            return await step.beginDialog('/auth');
+          } else {
+            return await step.next(step.options);
+          }
+        },
 
-      async step => {
-        const data = await service.getFaqBySubjectArray(min.instance.instanceId, 'faq', undefined);
-        const locale = step.context.activity.locale;
-        if (data !== undefined) {
-          await min.conversationalService.sendEvent(min, step, 'play', {
-            playerType: 'bullet',
-            data: data.slice(0, 10)
-          });
+        async step => {
+          const data = await service.getFaqBySubjectArray(min.instance.instanceId, 'faq', undefined);
+          const locale = step.context.activity.locale;
+          if (data !== undefined) {
+            await min.conversationalService.sendEvent(min, step, 'play', {
+              playerType: 'bullet',
+              data: data.slice(0, 10)
+            });
 
-          await min.conversationalService.sendText(min, step, Messages[locale].see_faq);
+            await min.conversationalService.sendText(min, step, Messages[locale].see_faq);
 
-          return await step.next();
+            return await step.next();
+          }
         }
-      }
-    ]));
+      ])
+    );
   }
 }
