@@ -1,32 +1,27 @@
 export const Messages = {
   'en-US': {
-    did_not_find: 'I\'m sorry I didn\'t find anything.',
+    did_not_find: "I'm sorry I didn't find anything.",
     going_answer: 'Great choice, now looking for your answer...',
-    wider_answer: subjectText =>
-      `Answering to you in a broader way... Not just about  ${subjectText}.`,
-    which_question: 'What\'s your question?',
+    wider_answer: subjectText => `Answering to you in a broader way... Not just about  ${subjectText}.`,
+    which_question: "What's your question?",
     anything_else: 'Can I help you with anything else?',
     here_is_subjects: 'Here are some subjects to choose from...',
     menu_select: 'Select',
-    lets_search: query =>
-      `Lets search for ${query}... What do you want to know?`,
-    see_faq: 'Please take a look at the FAQ I\'ve prepared for you. You can click on them to get the answer.',
+    lets_search: query => `Lets search for ${query}... What do you want to know?`,
+    see_faq: "Please take a look at the FAQ I've prepared for you. You can click on them to get the answer.",
 
     ask_first_time: 'What are you looking for?'
   },
   'pt-BR': {
     did_not_find: 'Desculpe-me, não encontrei nada a respeito.',
     going_answer: 'Ótima escolha, procurando resposta para sua questão...',
-    wider_answer: subjectText =>
-      `Vou te responder de modo mais abrangente... Não apenas sobre ${subjectText}`,
+    wider_answer: subjectText => `Vou te responder de modo mais abrangente... Não apenas sobre ${subjectText}`,
     which_question: 'Qual a pergunta?',
     anything_else: 'Posso ajudar em algo mais?',
     here_is_subjects: 'Aqui estão algumas categorias de assuntos...',
     menu_select: 'Selecionar',
-    lets_search: query =>
-      `Let's search about ${query}... What do you want to know?`,
-    see_faq:
-      'Veja algumas perguntas mais frequentes logo na tela. Clique numa delas para eu responder.',
+    lets_search: query => `Let's search about ${query}... What do you want to know?`,
+    see_faq: 'Veja algumas perguntas mais frequentes logo na tela. Clique numa delas para eu responder.',
 
     ask_first_time: 'Como eu posso ajudar?'
   }
