@@ -38,13 +38,13 @@
 
 import { TokenResponse } from 'botbuilder';
 import { GBLog, GBMinInstance, IGBDialog } from 'botlib';
-import { Messages } from '../strings';
+import { Messages } from '../strings.js';
 
 /**
  * Dialogs for handling Menu control.
  */
 export class OAuthDialog extends IGBDialog {
-  public static getOAuthDialog(min: GBMinInstance) {
+  public static getOAuthDialog (min: GBMinInstance) {
     return {
       id: '/auth',
       waterfall: [
