@@ -36,21 +36,18 @@
 
 'use strict';
 
-const urlJoin = require('url-join');
-
 import { GBDialogStep, GBLog, GBMinInstance, IGBCoreService, IGBPackage } from 'botlib';
 import { Sequelize } from 'sequelize-typescript';
-import { OAuthDialog } from './dialogs/OAuthDialog';
-import { ProfileDialog } from './dialogs/ProfileDialog';
-import { GuaribasGroup, GuaribasUser, GuaribasUserGroup } from './models';
+import { OAuthDialog } from './dialogs/OAuthDialog.js';
+import { ProfileDialog } from './dialogs/ProfileDialog.js';
+import { GuaribasGroup, GuaribasUser, GuaribasUserGroup } from './models/index.js';
 
 /**
  * Package for the security module.
  */
 export class GBSecurityPackage implements IGBPackage {
   public sysPackages: IGBPackage[];
-  public async getDialogs(min: GBMinInstance) {
-
+  public async getDialogs (min: GBMinInstance) {
     const out = [
       ProfileDialog.getNameDialog(min),
       ProfileDialog.getEmailDialog(min),
@@ -62,25 +59,24 @@ export class GBSecurityPackage implements IGBPackage {
       out.push(OAuthDialog.getOAuthDialog(min));
     }
     return out;
-
   }
-  public async unloadPackage(core: IGBCoreService): Promise<void> {
+  public async unloadPackage (core: IGBCoreService): Promise<void> {
     GBLog.verbose(`unloadPackage called.`);
   }
-  public async loadBot(min: GBMinInstance): Promise<void> {
+  public async loadBot (min: GBMinInstance): Promise<void> {
     GBLog.verbose(`loadBot called.`);
   }
-  public async unloadBot(min: GBMinInstance): Promise<void> {
+  public async unloadBot (min: GBMinInstance): Promise<void> {
     GBLog.verbose(`unloadBot called.`);
   }
-  public async onNewSession(min: GBMinInstance, step: GBDialogStep): Promise<void> {
+  public async onNewSession (min: GBMinInstance, step: GBDialogStep): Promise<void> {
     GBLog.verbose(`onNewSession called.`);
   }
-  public async onExchangeData(min: GBMinInstance, kind: string, data: any) {
+  public async onExchangeData (min: GBMinInstance, kind: string, data: any) {
     GBLog.verbose(`onExchangeData called.`);
   }
 
-  public async loadPackage(core: IGBCoreService, sequelize: Sequelize): Promise<void> {
+  public async loadPackage (core: IGBCoreService, sequelize: Sequelize): Promise<void> {
     core.sequelize.addModels([GuaribasGroup, GuaribasUser, GuaribasUserGroup]);
   }
 }
